cluster: Rename pacakge from clustering
