+++ internal/cluster/resource_cluster_group.go
@@ -1,4 +1,4 @@
-package clustering
+package cluster
 
 import (
 	"context"
