+++ internal/cluster/resource_cluster_group_member.go
@@ -1,4 +1,4 @@
-package clustering
+package cluster
 
 import (
 	"context"
