+++ internal/cluster/resource_cluster_group_member_test.go
@@ -1,4 +1,4 @@
-package clustering_test
+package cluster_test
 
 import (
 	"fmt"
