+++ internal/provider/provider.go
@@ -17,7 +17,7 @@ import (
 	incus_shared "github.com/lxc/incus/v6/shared/util"
 
 	"github.com/lxc/terraform-provider-incus/internal/certificate"
-	"github.com/lxc/terraform-provider-incus/internal/clustering"
+	"github.com/lxc/terraform-provider-incus/internal/cluster"
 	"github.com/lxc/terraform-provider-incus/internal/image"
 	"github.com/lxc/terraform-provider-incus/internal/instance"
 	"github.com/lxc/terraform-provider-incus/internal/network"
@@ -271,8 +271,8 @@ func (p *IncusProvider) Configure(ctx context.Context, req provider.ConfigureReq
 func (p *IncusProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		certificate.NewCertificateResource,
-		clustering.NewClusterGroupMemberResource,
-		clustering.NewClusterGroupResource,
+		cluster.NewClusterGroupMemberResource,
+		cluster.NewClusterGroupResource,
 		image.NewImageResource,
 		instance.NewInstanceResource,
 		instance.NewInstanceSnapshotResource,
@@ -297,7 +297,7 @@ func (p *IncusProvider) Resources(_ context.Context) []func() resource.Resource 
 
 func (p *IncusProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	dataSources := []func() datasource.DataSource{
-		clustering.NewClusterDataSource,
+		cluster.NewClusterDataSource,
 		image.NewImageDataSource,
 	}
